quiz revamp
