+++ front-end/src/quizes/payquiz/Quizparent.js
@@ -1,5 +1,19 @@
 import React, { useState, useEffect } from "react";
+import { makeStyles } from "@material-ui/core/styles";
 import { Quiz } from "./Quiz";
+import "./Quiz.css";
+
+const useStyles = makeStyles((theme) => ({
+    formControl: {
+        margin: theme.spacing(5),
+    },
+    fomrTopSpace: {
+        marginTop: "5vh",
+    },
+    button: {
+        margin: theme.spacing(1, 1, 0, 0),
+    },
+}));
 
 export const Quizparent = ({
     match: {
@@ -7,9 +21,8 @@ export const Quizparent = ({
     },
     history,
 }) => {
-    // STATE DECLARATIONS
-    // this state is used to store the quiz and answers object
-    const [state, setState] = useState([]);
+    const classes = useStyles();
+    const [state, setState] = useState(false);
     // this state is used to track the currently selected answer choice on the quiz
     const [value, setValue] = useState("");
     // this state is used to display an error if the selecte answer choice is incorrect
@@ -17,15 +30,7 @@ export const Quizparent = ({
     // this state is used to display helper text for correct and incorrect answers
     const [helperText, setHelperText] = useState("Buena Suerte!");
     // this state is used to set the current question and answer choices
-    const [question, setQuestion] = useState({
-        currentQuestion: "",
-        question: "",
-        Option1: "",
-        Option2: "",
-        Option3: "",
-        Option4: "",
-        questionId: "",
-    });
+    const [question, setQuestion] = useState(0);
 
     // this state keeps track of current user response(whether correcr or incorrect)
     // and is used to persist to server
@@ -43,6 +48,14 @@ export const Quizparent = ({
         score: 0,
     });
 
+    // fisherYalesShuffle
+    function shuffle(arr) {
+        for (let i = arr.length - 1; i > 0; i--) {
+            const j = Math.floor(Math.random() * (i + 1));
+            [arr[i], arr[j]] = [arr[j], arr[i]];
+        }
+        return arr;
+    }
     // Load the quiz into the state
     useEffect(() => {
         const payLoad = {
@@ -55,20 +68,12 @@ export const Quizparent = ({
         fetch(`http://localhost:3000/tests/${id}`, payLoad)
             .then((r) => r.json())
             .then((quizQuestionsObj) => {
+                for (let question of quizQuestionsObj) {
+                    question.answers = shuffle(question.answers);
+                }
                 setState(quizQuestionsObj);
-                // set the first question with default values
-                setQuestion({
-                    currentQuestion: 0,
-                    question: quizQuestionsObj[0].question,
-                    Option1: quizQuestionsObj[0].answers[0],
-                    Option2: quizQuestionsObj[0].answers[1],
-                    Option3: quizQuestionsObj[0].answers[2],
-                    Option4: quizQuestionsObj[0].answers[3],
-                    questionId: quizQuestionsObj[0].id,
-                });
             });
     }, []);
-
 
     // function to keep track of radio values
     const handleRadioChange = (event) => {
@@ -78,27 +83,9 @@ export const Quizparent = ({
     };
     // helper function to get currently selected answer and set response state
     const getSelectedRadio = (e) => {
-        let radioSelection = e.target.value;
-        const radioSelected = state[question.currentQuestion].answers.find(
-            (ans) => {
-                return ans.answer === radioSelection;
-            }
-        );
-        setResponse((prevState) => {
-            return {
-                ...prevState,
-                choice_id: radioSelected.id,
-                question_id: state[question.currentQuestion].id,
-            };
-        });
-    };
-
-    // retrive current question from state and set it
-    const handleQuestionChange = () => {
-        return (
-            state[question.currentQuestion] &&
-            state[question.currentQuestion].question
-        );
+        const currentValue = e.target.value;
+        const choice = state[question].answers.find((ans) => ans.answer === currentValue)
+        setResponse({ choice_id: choice.id, question_id: state[question].id })
     };
 
     // function to update local incorrect question tracking state
@@ -122,7 +109,7 @@ export const Quizparent = ({
     };
     // function to persist a correct answer to server
     const persistTrueResponse = () => {
-        const selectedResponse = {
+        const trueResponse = {
             user_id: localStorage.user,
             question_id: response.question_id,
             choice_id: response.choice_id,
@@ -134,7 +121,7 @@ export const Quizparent = ({
                 "Content-Type": "application/json",
                 Authorization: `Bearer ${localStorage.token}`,
             },
-            body: JSON.stringify(selectedResponse),
+            body: JSON.stringify(trueResponse),
         };
         fetch("http://localhost:3000/user_answers", answerPayLoad)
             .then((r) => r.json())
@@ -168,12 +155,13 @@ export const Quizparent = ({
 
     // persist total quiz results,only run at the end of quiz
     const persistResults = () => {
+        const testScore = results.no_correct - results.no_incorrect;
         const resultObj = {
-            ...results,
-            test_id: state[0].test.id,
-            score: results.no_correct - results.no_incorrect,
+            no_correct: testScore,
+            no_incorrect: results.no_correct,
+            test_id: state[question].test.id,
+            score: testScore,
         };
-
         const createPayload = {
             method: "POST",
             headers: {
@@ -191,65 +179,49 @@ export const Quizparent = ({
     };
 
     // a method to randomize questions for this approach
-    function shuffle(a) {
-        for (let i = a.length - 1; i > 0; i--) {
-            const j = Math.floor(Math.random() * (i + 1));
-            [a[i], a[j]] = [a[j], a[i]];
-        }
-        return a;
+    const isCorrect = (ans) => {
+        return state[question].answers.find((ans) => ans.is_correct === true)
     }
     // main function that combines all quiz logic based on previous questions
     // and whether the answer is correct and incorrect
     const handleSubmit = (event) => {
         event.preventDefault();
-        if (question.currentQuestion + 1 !== state.length) {
-            let comparison = state[question.currentQuestion].answers.find((ans) => {
-                if (ans.is_correct !== undefined && ans.is_correct === true) {
-                    return ans;
-                } else {
-                    setHelperText("Please select an option.");
-                    setError(true);
-                }
-            });
-            if (comparison.answer && comparison.answer === value) {
-                persistTrueResponse();
-                let arr = shuffle([0, 1, 2, 3]);
-                setQuestion((prevState) => {
-                    return {
-                        currentQuestion: prevState.currentQuestion + 1,
-                        question: state[prevState.currentQuestion + 1].question,
-                        Option1: state[prevState.currentQuestion + 1].answers[arr[0]],
-                        Option2: state[prevState.currentQuestion + 1].answers[arr[1]],
-                        Option3: state[prevState.currentQuestion + 1].answers[arr[2]],
-                        Option4: state[prevState.currentQuestion + 1].answers[arr[3]],
-                    };
-                });
-
+        if (question < state.length - 1) {
+            const correct = state[question].answers.find((ans) => ans.is_correct).id === response.choice_id;
+            if (correct) {
                 setHelperText("Buen Trabajo!");
-                setError(false);
                 setCorrectResults();
                 persistTrueResponse();
-            } else if (comparison.answer && comparison.answer !== value) {
+                const nextQuestion = question + 1;
+                setQuestion(nextQuestion)
+                setHelperText("Por favor, seleccione una opcion:");
+            } else if (!correct) {
                 setHelperText("Lo Siento, Intente Otra vez!");
                 setError(true);
                 setIncorrectResults();
                 persistFalseResponse();
-            } else {
-                setHelperText("Por favor, seleccione una opcion:");
-                setError(true);
             }
+            setError(false);
+            // setHelperText("Please select an option.");
         } else {
             setHelperText(
                 `Terminaste!Felicidades!, tuviste ${results.no_correct + 1 - results.no_incorrect
                 } preguntas correctas!`
             );
-        }
-        if (question.currentQuestion + 1 === state.length) {
-            persistResults();
+            setTimeout(persistResults, 2000);
         }
     }
-    return (<>
-        <Quiz />
-    </>)
 
+    return (
+        <Quiz
+            fetchedQuiz={state}
+            question={question}
+            error={error}
+            value={value}
+            handleRadioChange={handleRadioChange}
+            getSelectedRadio={getSelectedRadio}
+            helperText={helperText}
+            handleSubmit={handleSubmit}
+            history={history} />
+    )
 }
